camera-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
opencv-python-headless==4.9.0.80
insightface==0.7.3
onnxruntime==1.24.1
numpy==2.4.3
httpx==0.27.0
python-dotenv==1.0.1
psycopg2-binary==2.9.11
python-multipart==0.0.9
